Code refactor and giving descriptive names to classes
