+++ app/static/poll.js
@@ -89,4 +89,3 @@ function makeGraph(results) {
         }
     });
 }
-#ciao
