+++ app/templates/show_histogram.html
@@ -19,7 +19,7 @@
             
             <div class="card">
                 <img class="large-front-thumbnail"
-                src="{{ url_for('static', filename='histoimage/'+histogram_image  )}} "
+                src="{{ url_for('static', filename='histogram_image/'+histogram_image  )}} "
                 alt={{ image_id }}/>
                 
             </div>
