+++ ml/classification_utils.py
@@ -87,5 +87,5 @@ def classify_image(model_id, img_id):
     output = [(labels[idx], percentage[idx].item()) for idx in indices[0][:5]]
 
     img.close()
-    time.sleep(5)
+    time.sleep(2)
     return output
